move modal to root navigation
